chore: move SetupTestPrivSigner to testutil and avoid circular dependency
